Moved Format and Protocol into /adapter
